Added Favicon and Email hyperlink
Added new a transparent crest image to the imgs folder and made this the new favicon. Didn't know how to do this with the schema.js file so just added a header to each of the .html files mb. Also made it so you can click on the little email logo and it works as a hyperlink